+++ src/components/ui/calendar/calendar.tsx
@@ -3,8 +3,7 @@ import { useEffect, useState } from "react";
 import { leftPad } from "../../../lib/utils";
 import CalendarDay from "./calendar-day";
 import CalendarTimePopover from "./calendar-time-popover";
-import { ChevronLeftIcon, ChevronRightIcon } from "lucide-react";
-import { getWeek, add, startOfWeek, format, getYear } from "date-fns";
+import { add, startOfWeek, format } from "date-fns";
 import { nb } from "date-fns/locale/nb";
 import {
   Select,
@@ -15,62 +14,11 @@ import {
 } from "@/components/ui/select";
 import CalendarDayView from "./calendar-day-view";
 import CalendarWeekView from "./calendar-week-view";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import { Button } from "../button";
-import { DateTimePicker } from "./date-time-picker";
-import { setHours, setMinutes } from "date-fns";
-const weekDays = [
-  "Mandag",
-  "Tirsdag",
-  "Onsdag",
-  "Torsdag",
-  "Fredag",
-  "Lørdag",
-  "Søndag",
-];
+import PossibleBooking from "./booking-prompt";
 
-interface RelativeMousePositionProps {
-  x: number;
-  y: number;
-  day: number;
-  hours: number;
-  minutes: number;
-}
-function getDateAtMousePosition<RelativeMousePositionProps>(
-  e: React.MouseEvent<HTMLDivElement>,
-) {
-  const yPos = e.pageY - e.currentTarget.offsetTop;
-  const xPos = e.pageX - e.currentTarget.offsetLeft;
-  const height = e.currentTarget.clientHeight;
-  const width = e.currentTarget.clientWidth;
-  const day = Math.floor((xPos / width) * 7);
-  const time = Math.floor((yPos / height) * 288);
-  return {
-    x: xPos,
-    y: yPos,
-    day,
-    hours: Math.floor(time / 12),
-    minutes: (time % 12) * 5,
-  };
-}
-
-function compareMousePositions(
-  a: RelativeMousePositionProps | null,
-  b: RelativeMousePositionProps | null,
-) {
-  if (!a || !b) return false;
-  return (
-    a.x === b.x &&
-    a.y === b.y &&
-    a.day === b.day &&
-    a.hours === b.hours &&
-    a.minutes === b.minutes
-  );
-}
+import { setHours, setMinutes, toDate } from "date-fns";
+import {weekDays, RelativeMousePositionProps, getDateAtMousePosition, compareMousePositions} from "@/lib/utils";
+import PossiblePlaceholder from "./placeholder";
 
 export default function Calendar() {
   const [relativeMousePosition, setRelativeMousePosition] =
@@ -87,7 +35,6 @@ export default function Calendar() {
   const [endDate, setEndDate] = useState<Date | null>(null);
 
   useEffect(() => {
-    console.log(window.innerWidth);
     if (window.innerWidth < 900) {
       setView("day");
     }
@@ -145,66 +92,7 @@ export default function Calendar() {
     }
   }
 
-  function PossiblePlaceholder() {
-    return (
-      !dragEnd &&
-      dragStart &&
-      relativeMousePosition && (
-        <div
-          className="rounded-md absolute bg-neutral-300 shadow-sm border z-10"
-          style={{
-            top: dragStart.y,
-            left: view == "week" ? dragStart.day * (100 / 7) + "%" : "0",
 
-            height: relativeMousePosition.y - dragStart.y,
-            width: view == "week" ? 100 / 7 + "%" : "100%",
-          }}
-        ></div>
-      )
-    );
-  }
-
-  function PossibleBooking() {
-    return (
-      dragEnd &&
-      dragStart && (
-        <div
-          className="rounded-md absolute bg-neutral-400 shadow-sm border z-10"
-          style={{
-            top: dragStart.y,
-            left: view == "week" ? dragStart.day * (100 / 7) + "%" : "0",
-
-            height: dragEnd.y - dragStart.y,
-            width: view == "week" ? 100 / 7 + "%" : "100%",
-          }}
-        >
-          <div className="relative h-full w-full">
-            <div
-              className="absolute top-full left-0 p-4 border rounded-md w-64 bg-white shadow-sm"
-              onMouseDown={(e) => e.stopPropagation()}
-              onMouseUp={(e) => e.stopPropagation()}
-              onMouseMove={(e) => {
-                e.stopPropagation();
-                setRelativeMousePosition(null);
-              }}
-            >
-              <div className="text-sm text-neutral-500">Start:</div>
-              <DateTimePicker
-                date={startDate || new Date()}
-                setDate={setStartDate}
-              />
-              <div className="text-sm text-neutral-500">Slutt:</div>
-              <DateTimePicker
-                date={endDate || new Date()}
-                setDate={setEndDate}
-              />
-              <Button className="w-full">Reserver</Button>
-            </div>
-          </div>
-        </div>
-      )
-    );
-  }
 
   return (
     <div className="w-full h-[200vh] max-w-7xl flex flex-col mx-auto cursor-pointer select-none">
@@ -277,8 +165,21 @@ export default function Calendar() {
             }
           />
         )}
-        <PossiblePlaceholder />
-        <PossibleBooking />
+        <PossiblePlaceholder dragStart={dragStart!} view={view} relativeMousePosition={relativeMousePosition!} visible={
+      (!dragEnd &&
+      dragStart &&
+      relativeMousePosition) as boolean
+          } />
+        <PossibleBooking
+        startDate={startDate}
+        endDate={endDate}
+        dragStart={dragStart!}
+        dragEnd={dragEnd!}
+        setRelativeMousePosition={setRelativeMousePosition}
+        setEndDate={setEndDate}
+        setStartDate={setStartDate} 
+        view={view}
+        />
         {view == "week" ? (
           weekDays.map((_, index) => <CalendarDay key={index} index={index} />)
         ) : (
